backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
structlog==24.1.0
tenacity==8.2.3
greenlet==3.0.3
